Revert "refactor(app): share workspace picker popover"
This reverts commit 1404c97a775cc071ab01488911f364a68fc032dc.
